Creating an executable file to run all check exercise files using a file path
